+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add default settings for user status and preferred contact in new user creation. Refs UIU-1385.
 * Fix navigation paths on cancel button click on loan details, open and closed loan lists pages. Refs UIU-1377.
 * Prevent loan details opening upon click in loans action menu without selecting link. Fixes UIU-1359.
+* Display correct user block on the edit form. Fixes UIU-1397.
 
 ## [2.26.0](https://github.com/folio-org/ui-users/tree/v2.26.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.3...v2.26.0)