+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-users
 
+## 2.26.1 (IN PROGRESS)
+
+* Correctly handle checkboxes in MCLs. Refs UIU-1408.
+
 ## [2.26.0](https://github.com/folio-org/ui-users/tree/v2.26.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.3...v2.26.0)